+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.63.2
+
 ### Fixed
 
 - Automatic Persisted Queries work with hybrid cache mode https://github.com/nuwave/lighthouse/pull/2727